requirements.txt
kivy==2.2.1
kivymd==1.1.1
plyer==2.1.0
android-permissions==1.1.1
cryptography==41.0.1
pillow==10.0.0
pynput==1.7.6
schedule==1.2.0
buildozer==1.5.0 